requirements.txt
Django==4.0.6
pyTelegramBotAPI==4.6.0
psycopg2==2.9.3

python-decouple==3.6
Pillow==9.2.0

emoji==2.0.0